增加管理员删除报修

+++ code/elife/lifeservice-server/src/main/java/lifeservice/lifeservice/controller/MaintainController.java
@@ -175,8 +175,13 @@ public class MaintainController {
 
     @RequestMapping(path = "/deleteMaintain")
     @ResponseBody
-    public boolean deleteMaintain(@RequestParam long id){
-
+    public boolean deleteMaintain(HttpServletRequest request, @RequestParam long id){
+        HttpSession session = request.getSession();
+        String name = (String) session.getAttribute("username");
+        String role = (String) session.getAttribute("role");
+        if (StringUtils.isEmpty(name) || !("1".equals(role))) {
+            return false;
+        }
         return maintainService.deleteById(id);
     }
 }
